refactor: reorganize interface imports for consistency in core library
