+++ project/libs/core/src/index.ts
@@ -1,7 +1,7 @@
 export { Entity } from './lib/classes/entity';
 
-export { UserRole } from './lib/types/user/user-role.type';
-export { User } from './lib/types/user/user.interface';
+export { UserRole } from './lib/interfaces/user/user-role.type';
+export { User } from './lib/interfaces/user/user.interface';
 
-export { StorableEntity } from './lib/interfaces/storable-entity.interface';
-export { EntityFactory } from './lib/interfaces/entity-factory.interface';
+export { StorableEntity } from './lib/interfaces/base/storable-entity.interface';
+export { EntityFactory } from './lib/interfaces/base/entity-factory.interface';
